perf: 修改作业模板的全局变量默认值，希望同步到执行方案 #3954
1. 抽取TaskPlanServiceImpl中变量跟随作业的逻辑
2. TaskPlanInfoDao中buildPlanInfo方法太长，重构
