+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/service/plan/TaskPlanService.java
@@ -310,10 +310,4 @@ public interface TaskPlanService {
 
     boolean batchUpdatePlanVariable(List<TaskPlanInfoDTO> planInfoList);
 
-    /**
-     * 更新执行方案的版本号
-     * 作业执行方案的变量是否跟随作业模板变量，如果是，且执行方案变量默认值跟作业模板的不一样，
-     * 该执行方案生成新版本号，产生差异，后续可以去同步
-     */
-    void updatePlanVersionIfVarValueChanged(TaskTemplateInfoDTO taskTemplateInfo, Boolean changed);
 }
